requirements.txt
torch
torchvision
transformers
datasets
numpy
matplotlib
pyyaml
beautifulsoup4
tqdm
pillow

